requirements.txt
pynput
Pillow
psutil
requests
